requirements.txt
lxml==4.3.4
requests==2.22.0
beautifulsoup4==4.7.1
symspellpy==6.3.8
pandas==0.24.2
datetime==4.3
cld2-cffi==0.1.4
nltk==3.4.4
